Move gridGallery and add icons and header
