+++ web-app/src/components/GridGallery/GridGalleryContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { photos } from '../photos';
+import { photos } from '../../photos';
 import { GridGallery } from './GridGallery';
 
 export class GridGalleryContainer extends React.Component {
